Fixed Ui Glitches, Made drawer full Screen, Added Settings button on the top app bar
