+++ app/build.gradle.kts
@@ -82,6 +82,7 @@ dependencies {
     implementation("androidx.constraintlayout:constraintlayout:2.1.4")
     implementation("androidx.navigation:navigation-runtime-ktx:2.8.5")
     implementation("androidx.navigation:navigation-compose:2.8.5")
+    implementation("androidx.games:games-activity:3.0.5")
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.2.1")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.6.1")
